+++ src/api/volume.rs
@@ -3,9 +3,10 @@
 use crate::{
     conn::{Headers, Payload},
     models,
-    opts::{VolumeCreateOpts, VolumeListOpts, VolumePruneOpts},
+    opts::{ClusterVolumeUpdateOpts, VolumeCreateOpts, VolumeListOpts, VolumePruneOpts},
     Result,
 };
+use containers_api::url;
 
 impl_api_ty!(Volume => name);
 
@@ -14,6 +15,15 @@ impl Volume {
         Inspect -> &format!("/volumes/{}", vol.name), models::Volume
         Delete -> &format!("/volumes/{}", vol.name), ()
     }
+
+    api_doc! { Volume => Update
+    |
+    /// Update a volume. Valid only for Swarm cluster volumes
+    pub async fn update(&self, opts: &ClusterVolumeUpdateOpts) -> Result<()> {
+        let mut ep = format!("/volumes/{}", self.name());
+        url::append_query(&mut ep, url::encoded_pair("version", opts.version()));
+        self.docker.put(&ep, Payload::Json(opts.serialize()?)).await.map(|_| ())
+    }}
 }
 
 impl Volumes {
@@ -25,7 +35,7 @@ impl Volumes {
     |
     /// List available volumes
     pub async fn list(&self, opts: &VolumeListOpts) -> Result<models::VolumeListResponse> {
-        let ep = containers_api::url::construct_ep("/volumes", opts.serialize());
+        let ep = url::construct_ep("/volumes", opts.serialize());
         self.docker.get_json(&ep).await
     }}